think i got it
